Removed debug spheres.
Tried to adjust skeleton attack sprite but could not edit it in Editor since it's not at the same grid size as the other sprites. Its origin is slightly at different location compared to other sprites. Should edit it in a different software but I don't want to spend that time on this free asset that I won't use.